pyproject.toml
[project]
name = "prawobiorca-backend"
version = "0.17.0"
description = "Backend for law acts search engine with AI features."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg==0.30.*",
    "bcrypt==4.3.*",
    "fastapi==0.115.*",
    "google-cloud-storage==3.1.*",
    "granian==2.2.*",
    "pydantic-settings==2.9.*",
    "python-multipart==0.0.20",
    "qdrant-client==1.14.*",
    "redis==6.2.*",
    "sqlalchemy[asyncio]==2.0.*",
]

[project.optional-dependencies]
dev = [
    "alembic==1.16.*",
    "bandit==1.8.*",
    "black==25.1.*",
    "granian[reload]==2.2.*",
    "isort==6.0.*",
    "psycopg==3.2.*",
    "pytest==8.3.*",
    "pytest-asyncio==1.0.*",
    "uvicorn==0.34.*",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.black]
line-length = 120
target-version = ["py313"]
exclude = '''
/(
    \.git
  | \.venv
  | config
  | secrets
  | bash
  | alembic
)/
'''

[tool.isort]
skip = ["alembic", ".venv"]
